+++ keyboards/handwired/horrortroll/k552/k552.c
@@ -15,3 +15,19 @@
  */
 
 #include "k552.h"
+
+led_config_t g_led_config = { {
+    { }
+}, {
+    {152,  0}, {165,  0}, {190,  0}, {205,  0},
+    {224, 21}, {224, 43}, {224, 54},
+    {188, 64}, {172, 64}, {156, 64}, {140, 64}, {115, 64}, {99 , 64}, {75 , 64}, {59 , 64}, {43 , 64}, {26 , 64},
+    {0  , 15}, {0  , 50}, {0  , 39},
+    {18 ,  0}, {36 ,  0}, {57 ,  0}, {67 ,  0}
+}, {
+    2, 2, 2, 2,       2, 2, 2, 2,
+    2,                         2,
+    2,                         2,
+    2,                         2,
+    2, 2, 2, 2, 2, 2, 2, 2, 2, 2,
+} };
