--- keyboards/handwired/horrortroll/k552/lib/layers.c
@@ -1,14 +1,0 @@
-#pragma once
-
-// Each layer gets a name for readability, which is then used in the keymap matrix below.
-// The underscores don't mean anything - you can have a layer called STUFF or any other name.
-// Layer names don't all need to be of the same length, obviously, and you can also skip them
-// entirely and just use numbers.
-
-enum layer_names {
-    _BASE = 0,
-    _WAVE = 1,
-    _FN = 2
-};
-
-bool process_record_user_oled(uint16_t keycode, keyrecord_t *record);
