+++ keyboards/handwired/horrortroll/k552/rules.mk
@@ -27,4 +27,7 @@ OLED_ENABLE = yes
 OLED_DRIVER = SSD1306
 WPM_ENABLE = yes
 
+RGB_MATRIX_ENABLE = yes
+RGB_MATRIX_DRIVER = WS2812
+
 LAYOUTS = tkl_ansi
